Network Tokens support for upf-epc
Network tokens are an open and secure method for end users and application
providers to coordinate with the network about how their traffic is treated
(e.g., to access a 5G slice, a firewall whitelist, or a zero-rating service).
They replace complex, insecure, and privacy-invasive DPI application signatures
with a deterministic and unified mechanism to expose datapath services.

Build the image with `--build-arg ENABLE_NTF=1` to enable NTF.