QoS: Add support for metering (#310)
This commit adds QoS module which can be thought of as an ExactMatch
module, with metering action using DPDK metering library. Arguments
to program the metering in the module - cir, pir, cbs, pbs and ebs are
exactly same as defined in the DPDK library.

This module has 4 reserved ogates -
0 - METER - To indicate to modules that traffic matching entry with this
            gate need to be metered. Based on the result they will be
            sent out of one of the following gates
1 - GREEN - For traffic conformant to CIR
2 - YELLOW - For traffic conformant to PIR
3 - RED   - For traffic exceeding PIR

Apart from these there 3 more user-defined gates in up4.bess -
4 - Default gate for lookup failures
5 - Explicitly asked to drop, modelling "Gate Status" closed
6 - Unmetered traffic, modelling PDRs without QER ID

metering.bess unit-test is provided to represent the UPF usage.

PFCP agent creates 2 entries for every incoming QER entry, as it consists
of UL and DL but the module is agnostic to the direction. GBR maps to CIR
and MBR maps to PIR.

Fixes: #299

Co-authored-by: Badhrinath Padmanabhan <badhrinath@opennetworking.org>
Co-authored-by: Amar Srivastava <amar.srivastava@intel.com>
Co-authored-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>

Co-authored-by: badhri85 <badhri85@node.onf5.cord-testdrive-pg0.wisc.cloudlab.us>
Co-authored-by: Badhrinath Padmanabhan <badhrinath@opennetworking.org>
Co-authored-by: Amar Srivastava <amar.srivastava@intel.com>
Co-authored-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>