QoS: Use params to profile map, carry pointer
DPDK split the meter and profile config objects because in practice they
found a small number of latter. So instead of repeating it for every
entry we now only carry a pointer to the profile. The profile itself is
stored in params to profile map. Whenever we get a commandAdd, we lookup
if a profile was created using the same params. If found we use copy
it's address, else we insert and do the same.

Signed-off-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>