Fix flow measure reads in no traffic situations (#400)
When there is no traffic flowing though the pipeline, the in-band mechanism of attaching the current buffer side to packets can't be used to signal buffer changes to follower modules. Instead, we have to flip the flag on the leader and explicitly tell the module which side to read from on the controller/agent side.