Export Prometheus flow metrics with UE IP labels (#375)
This PR enriches the flow metrics exported by adding an UE IP address label. The IP is looked up by searching the PDRs of a given sessions for the N6 uplink rule, which contains the UE IP as the source IP.
This is implemented by adding a `PfcpNodeCollector` which makes the `PfcpNode` Prometheus observable, similar to the `upfCollector`.