Fix protoc `go install` command in Dockerfile (#580)
This is the smallest possible quick-fix to address the failing protoc `go get` command in the Dockerfile. No other changes needed, as it preserves the exact behavior as of now, for better or worse. This just unblocks other PRs, further improvements can be deferred.

Follows official guidance: https://go.dev/doc/go-get-install-deprecation